rename expose command to start
