+++ lib/start/utils/replace_assets_host.js
