+++ lib/start/utils/sync_files.js
